# backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ContractRequest": {
				"properties": {
					"payout": {
						"type": "string"
					},
					"choice": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payout",
					"choice"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RaceStatus": {
				"enum": [
					"PENDING",
					"IN_PROGRESS",
					"FINISHED"
				],
				"type": "string"
			},
			"Dog": {
				"properties": {
					"name": {
						"type": "string"
					},
					"number": {
						"type": "number",
						"enum": [
							1,
							2,
							3,
							4
						]
					}
				},
				"required": [
					"name",
					"number"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Choice": {
				"properties": {
					"id": {
						"type": "string"
					},
					"payout": {
						"type": "string"
					},
					"dog": {
						"$ref": "#/components/schemas/Dog"
					},
					"outpoint": {
						"properties": {
							"vout": {
								"type": "number",
								"format": "double"
							},
							"txid": {
								"type": "string"
							}
						},
						"required": [
							"vout",
							"txid"
						],
						"type": "object"
					},
					"parity": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"payout",
					"dog",
					"outpoint",
					"parity",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Race": {
				"properties": {
					"id": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/RaceStatus"
					},
					"available": {
						"items": {
							"$ref": "#/components/schemas/Dog"
						},
						"type": "array"
					},
					"chosen": {
						"items": {
							"$ref": "#/components/schemas/Choice"
						},
						"type": "array"
					},
					"winner": {
						"$ref": "#/components/schemas/Dog"
					},
					"houseWon": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"status",
					"available",
					"chosen",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IChoicePayload": {
				"properties": {
					"payout": {
						"type": "string"
					},
					"dog": {
						"$ref": "#/components/schemas/Dog"
					},
					"outpoint": {
						"properties": {
							"vout": {
								"type": "number",
								"format": "double"
							},
							"txid": {
								"type": "string"
							}
						},
						"required": [
							"vout",
							"txid"
						],
						"type": "object"
					},
					"parity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"payout",
					"dog",
					"outpoint",
					"parity"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@satoshi-dog/backend",
		"version": "0.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/contract": {
			"post": {
				"operationId": "GetContract",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Contract"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContractRequest"
							}
						}
					}
				}
			}
		},
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/races": {
			"post": {
				"operationId": "NewRace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Race"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": []
			},
			"get": {
				"operationId": "GetRaceIDs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": []
			}
		},
		"/races/{id}": {
			"delete": {
				"operationId": "DeleteRace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "UpdateRace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Race"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IChoicePayload"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetRace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Race"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/races/{id}/run": {
			"post": {
				"operationId": "RunRace",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"winnerTxID": {
											"type": "string"
										},
										"winnerDog": {
											"$ref": "#/components/schemas/Dog"
										},
										"houseWon": {
											"type": "boolean"
										}
									},
									"required": [
										"houseWon"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Race"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}